backup_before_migration/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests>=2.32.3
pydantic>=2.5.0
pydantic-settings>=2.0.3
clickup-python-sdk>=2.0.1
aiohttp==3.9.1
python-multipart==0.0.6
jinja2==3.1.2
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
langgraph>=0.2.35  # Requerido para sistema de logging automático
aiosqlite>=0.19.0  # Requerido para SQLite async (fallback)
asyncpg>=0.29.0  # Requerido para PostgreSQL async
aiosmtplib==3.0.1
httpx==0.25.0
twilio==8.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
reportlab==4.2.2
email-validator>=2.1.0  # Requerido para EmailStr de Pydantic
sentence-transformers>=2.2.0  # Requerido para el motor de búsqueda RAG
qrcode[pil]>=7.4.2  # Requerido para generar códigos QR de WhatsApp
Pillow>=10.0.0  # Requerido para procesamiento de imágenes de QR
psutil>=5.9.0  # Requerido para monitoreo del sistema Railway
